Rename org.lflang.lfc to org.lflang.cli
