+++ org.lflang.cli/build.gradle
@@ -31,7 +31,7 @@ apply plugin: 'com.github.johnrengelman.shadow'
 task buildLfc() {
     apply plugin: 'application'
     apply plugin: 'com.github.johnrengelman.shadow'
-    mainClassName = 'org.lflang.lfc.Main'
+    mainClassName = 'org.lflang.cli.Lfc'
 
     shadowJar {
         exclude 'test/*'
@@ -57,6 +57,6 @@ task runLfc(type: JavaExec) {
     description = "Build and run lfc, use --args to pass arguments"
     group = "application"
     classpath = sourceSets.main.runtimeClasspath
-    mainClass = 'org.lflang.lfc.Main'
+    mainClass = 'org.lflang.cli.Lfc'
     workingDir = '..'
 }
