+++ org.lflang.cli/src/org/lflang/cli/LFStandaloneModule.java
@@ -26,7 +26,7 @@
  * generated by Xtext 2.23.0
  */
 
-package org.lflang.lfc;
+package org.lflang.cli;
 
 import java.util.Objects;
