+++ org.lflang.cli/test/kotlin/org/lflang/cli/tests/LfcIssueReportingTest.kt
@@ -21,14 +21,14 @@
  * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF
  * THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.lflang.lfc.tests
+package org.lflang.cli.tests
 
 import junit.framework.Assert.assertEquals
 import junit.framework.AssertionFailedError
 import org.junit.Test
 import org.lflang.LFRuntimeModule
 import org.lflang.LFStandaloneSetup
-import org.lflang.lfc.*
+import org.lflang.cli.*
 import java.io.*
 import java.nio.charset.StandardCharsets
 import java.nio.file.Files
@@ -45,7 +45,7 @@ class SpyPrintStream {
 }
 
 
-class LfIssueReportingTest {
+class LfcIssueReportingTest {
     /*
         Note: when executing these tests in Intellij, I get the following error:
 
@@ -115,7 +115,7 @@ class LfIssueReportingTest {
         val backend = ReportingBackend(Io(err = stderr.ps), AnsiColors(useColors), 2)
         val injector = LFStandaloneSetup(LFRuntimeModule(), LFStandaloneModule(backend))
             .createInjectorAndDoEMFRegistration()
-        val main = injector.getInstance(Main::class.java)
+        val main = injector.getInstance(Lfc::class.java)
 
         val packageName = loader.packageName.replace('.', '/')
         // relative to root of gradle project
