+++ org.lflang.cli/test/resources/org/lflang/cli/tests/emptyFile.stderr
