+++ org.lflang.cli/test/resources/org/lflang/cli/tests/simpleWarning.lf
