+++ settings.gradle
@@ -1,3 +1,3 @@
 include 'org.lflang'
-include 'org.lflang.lfc'
+include 'org.lflang.cli'
 include 'org.lflang.tests'
